+++ c/binary-search/src/binary_search.h
@@ -1,4 +1,4 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int* binary_search(int target, int arr[], size_t length);
+int* binary_search(int_fast16_t target, int arr[], size_t length);
